requirements.txt
environs==14.2.0
requests==2.32.4
python-telegram-bot==13.0